First attempt at adding basic vbucket functionality to mutilate. This version ONLY supports one server, has no agent mode, and takes in a vbucket configuration through a file.
Since this does not yet include the auto-configuration updater thread
and thus works without locking the vBucket map, it should be the fastest
possible implementation.

A config.txt file is included that maps all vBuckets to zephyr5. To run,

./mutilate --server=zephyr5-10G:11210 -T 8 --binary --username=default
	--membaseConfig=config.txt

Note: omitting the last parameter will result in a test without using
vBuckets, which is equivalent to the previous commit of mutilate.